Split the kendo ui project's tests into their own project and fix some nuget references.
