+++ Bumblebee.Examples.Web.IntegrationTests/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="FluentAssertions" version="3.2.2" targetFramework="net45" />
-  <package id="NUnit" version="2.6.3" targetFramework="net45" />
+  <package id="NUnit" version="2.6.4" targetFramework="net45" />
   <package id="PhantomJS" version="1.9.8" targetFramework="net45" />
   <package id="Selenium.Support" version="2.44.0" targetFramework="net45" />
   <package id="Selenium.WebDriver" version="2.44.0" targetFramework="net45" />
