+++ Bumblebee.KendoUI.IntegrationTests/Bumblebee.KendoUI.IntegrationTests.csproj
@@ -2,13 +2,13 @@
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Release</Configuration>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{8EE25B3E-3903-4E14-BD1F-AB6836C7B364}</ProjectGuid>
+    <ProjectGuid>{74ADC4D7-9473-4978-9FCF-5C7F70A66BA2}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Bumblebee.KendoUI</RootNamespace>
-    <AssemblyName>Bumblebee.KendoUI</AssemblyName>
+    <RootNamespace>Bumblebee.KendoUI.IntegrationTests</RootNamespace>
+    <AssemblyName>Bumblebee.KendoUI.IntegrationTests</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\</SolutionDir>
@@ -32,32 +32,59 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="FluentAssertions">
+      <HintPath>..\packages\FluentAssertions.3.2.2\lib\net45\FluentAssertions.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <Reference Include="FluentAssertions.Core">
+      <HintPath>..\packages\FluentAssertions.3.2.2\lib\net45\FluentAssertions.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="nunit.framework">
+      <HintPath>..\packages\NUnit.2.6.4\lib\nunit.framework.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.XML" />
+    <Reference Include="System.Xml.Linq" />
     <Reference Include="WebDriver">
       <HintPath>..\packages\Selenium.WebDriver.2.44.0\lib\net40\WebDriver.dll</HintPath>
     </Reference>
+    <Reference Include="WebDriver.Support">
+      <HintPath>..\packages\Selenium.Support.2.44.0\lib\net40\WebDriver.Support.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="KendoDatePicker.cs" />
-    <Compile Include="KendoMultiSelectOption.cs" />
-    <Compile Include="KendoMultiSelect.cs" />
-    <Compile Include="KendoDropDownList.cs" />
-    <Compile Include="KendoDropDownListOption.cs" />
-    <Compile Include="KendoNumericTextBox.cs" />
+    <Compile Include="Given_kendo_date_picker.cs" />
+    <Compile Include="Given_kendo_drop_down_list_and_no_options_selected.cs" />
+    <Compile Include="Given_kendo_drop_down_list_when_clicking_an_option.cs" />
+    <Compile Include="Given_kendo_drop_down_list_with_multiple_values.cs" />
+    <Compile Include="Given_kendo_numeric_text_box_when_entering_value.cs" />
+    <Compile Include="Pages\KendoDatePickerDemoPage.cs" />
+    <Compile Include="Pages\KendoDropDownListDemoPage.cs" />
+    <Compile Include="Pages\KendoMultiSelectDemoPage.cs" />
+    <Compile Include="Pages\KendoNumericTextBoxDemoPage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="KendoExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="Bumblebee.KendoUI.nuspec" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Bumblebee.KendoUI\Bumblebee.KendoUI.csproj">
+      <Project>{8ee25b3e-3903-4e14-bd1f-ab6836c7b364}</Project>
+      <Name>Bumblebee.KendoUI</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Bumblebee\Bumblebee.csproj">
       <Project>{690961a6-7130-4afb-ae14-51d18dc6b223}</Project>
       <Name>Bumblebee</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="..\packages\Selenium.WebDriver.IEDriver.2.44.0.0\content\IEDriverServer.exe">
+      <Link>IEDriverServer.exe</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
@@ -66,6 +93,17 @@
       <ErrorText>This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
     </PropertyGroup>
     <Error Condition="!Exists('$(SolutionDir)\.nuget\NuGet.targets')" Text="$([System.String]::Format('$(ErrorText)', '$(SolutionDir)\.nuget\NuGet.targets'))" />
+  </Target>
+  <PropertyGroup Label="DownloadIEDriverBuildTask">
+    <CoreBuildDependsOn>DownloadIEDriver;$(CoreBuildDependsOn)</CoreBuildDependsOn>
+    <IEDriverInstallPath>$(SolutionDir)packages\Selenium.WebDriver.IEDriver.2.44.0.0\</IEDriverInstallPath>
+    <IEDriverToolsPath>$(IEDriverInstallPath)tools\</IEDriverToolsPath>
+    <IEDriverInitScriptPath>$(IEDriverToolsPath)Init.ps1</IEDriverInitScriptPath>
+    <IEDriverPath>$(IEDriverInstallPath)content\IEDriverServer.exe</IEDriverPath>
+    <PowerShellExe Condition="'$(PowerShellExe)'==''">%WINDIR%\System32\WindowsPowerShell\v1.0\powershell.exe</PowerShellExe>
+  </PropertyGroup>
+  <Target Name="DownloadIEDriver" Label="DownloadIEDriverBuildTask">
+    <Exec Condition="!Exists('$(IEDriverPath)')" Command="$(PowerShellExe) -NonInteractive -executionpolicy Unrestricted -command &quot;&amp; { &amp;'$(IEDriverInitScriptPath)' '$(IEDriverInstallPath)' '$(IEDriverToolsPath)'} &quot;" />
   </Target>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
