+++ Bumblebee.KendoUI.IntegrationTests/Given_kendo_drop_down_list_when_clicking_an_option.cs
@@ -1,12 +1,15 @@
 using System.Linq;
+
 using Bumblebee.Extensions;
-using Bumblebee.IntegrationTests.Shared.Pages.KendoUI;
+using Bumblebee.KendoUI.IntegrationTests.Pages;
 using Bumblebee.Setup;
 using Bumblebee.Setup.DriverEnvironments;
+
 using FluentAssertions;
+
 using NUnit.Framework;
 
-namespace Bumblebee.IntegrationTests.Bumblebee.KendoUI
+namespace Bumblebee.KendoUI.IntegrationTests
 {
     // ReSharper disable InconsistentNaming
     [TestFixture]
