+++ Bumblebee.KendoUI.IntegrationTests/packages.config
@@ -1,12 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="FluentAssertions" version="3.2.2" targetFramework="net45" />
-  <package id="Nancy" version="0.23.2" targetFramework="net45" />
-  <package id="Nancy.Hosting.Self" version="0.23.2" targetFramework="net45" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
-  <package id="PhantomJS" version="1.9.8" targetFramework="net45" />
   <package id="Selenium.Support" version="2.44.0" targetFramework="net45" />
   <package id="Selenium.WebDriver" version="2.44.0" targetFramework="net45" />
   <package id="Selenium.WebDriver.IEDriver" version="2.44.0.0" targetFramework="net45" />
-  <package id="WebDriver.ChromeDriver.win32" version="2.13.0.0" targetFramework="net45" />
 </packages>
